core: moving files around
